requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.7.3
coverage==4.5.2
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
Pillow==5.4.1
pluggy==0.9.0
py==1.8.0
pytest==4.3.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
